treewide: Fix code issues detected using GCC 8.2.0
Use the latest version of GCC to take advantage of improved static analysis.
These issues appeared as warnings from the compiler.

Many of these fixes are for clearly incorrect code; compiler warnings
should not be taken lightly.

+++ sound/soc/msm/qdsp6v2/msm-pcm-voice-v2.c
@@ -496,7 +496,7 @@ static int msm_voice_sidetone_put(struct snd_kcontrol *kcontrol,
 					struct snd_ctl_elem_value *ucontrol)
 {
 	int ret;
-	long sidetone_enable = ucontrol->value.integer.value[0];
+	int sidetone_enable = ucontrol->value.integer.value[0];
 	uint32_t session_id = ALL_SESSION_VSID;
 
 	ret = voc_set_afe_sidetone(session_id, sidetone_enable);
